UserProfileHelperService.uploadProfilePhoto added to simplify uploading profile photos
